+++ utils/queue/queue.go
@@ -5,7 +5,7 @@ import "container/heap"
 // An Item is something we manage in a Priority queue.
 type Item[T any] struct {
 	Value    T
-	Priority uint8
+	Priority int
 	// The index is needed by update and is maintained by the heap.Interface methods.
 	index int // The index of the item in the heap.
 }
@@ -14,7 +14,7 @@ type PriorityQueue[T any] struct {
 	ipq internalPriorityQueue[T]
 }
 
-func (pq *PriorityQueue[T]) Push(x T, priority uint8) {
+func (pq *PriorityQueue[T]) Push(x T, priority int) {
 	item := Item[T]{
 		Value:    x,
 		Priority: priority,
@@ -30,7 +30,7 @@ func (pq *PriorityQueue[T]) Pop() (T, bool) {
 	return heap.Pop(&pq.ipq).(Item[T]).Value, true
 }
 
-func (pq *PriorityQueue[T]) Update(item Item[T], value T, priority uint8) {
+func (pq *PriorityQueue[T]) Update(item Item[T], value T, priority int) {
 	item.Value = value
 	item.Priority = priority
 	heap.Fix(&pq.ipq, item.index)
